Rewrite of the model explanation.
The explanation of the computational model and how it relates to the
experimental observations was very very bad. Gallagher wanted a whole
rewrite.

After talking with Gallagher, I came down on an explanation based on the
three states "downhill" 7/6pi and 1/6pi and "uphill" 1/6pi all being
qualitatively the same, while "uphill" 7/6 pi changes meaningfully.

This is written out with figures showing orbits. Was that a success? We'll
see what Gallagher thinks.