Added notes for Intro and Background Sects
I find myself a bit confused when writing later sections about what is or
isn't covered in early sections.

I've added very brief and sloppy notes on what I think aught to be
discussed in the Introduction and Background sections. Hopefully this
helps my thought process and is a reference if I need to start on those
sections.