Edits for Reviewer
Tom made a change to the first paragraph to emphasize that the QM
picture is coherent while ours is not.

The reviewer noted one typo of W_0 that should be E_s, at the
first reference to Fig. Sim0.

In a few places, mostly Figs. DILP2, DILM14, DILM30, E_S was changed to
E_s for consistency.